refactor: add helper method for bulk secret creation
